Rename AddHeader to Telemetry policy
